Sets up 7-seg display and LPDDR RAM interface.
The seven segment display has 3 digits and will be used for displaying
status/error codes to the user for debugging. It currently tells if the
LPDDR RAM is done calibrating or has an error. I have also included the
UCF file which has the signals for LPDDR RAM, seven segment display,
and system clock with many other modules there but still commented. In
addition I have included a file called MimasV2Config.py, which is a script
for flashing the Mimas V2 on Linux. The LPDDR RAM interface module has
been generated by the Xilinx ISE software, but I have modified it slightly
so that the c3_sys_clk signal is not buffered. This was causing problems
when using the master clock both in the RAM and elsewhere because
apparently an input can't drive both buffered and unbuffered components.
Some minimal ISE project files are also included.