Adds rolling VGA buffer connected to main memory.
The VGA buffer continuously fills itself by reading the next line from main
memory that needs to be displayed by the VGA display controller. Each of these
lines is represented by 256 8-bit colors to construct the 256x192 Consolite
display. Since the VGA display controller outputs 640x480 pixels, each of
the Consolite pixels is either 2 or 3 VGA pixels wide and high, in an
alternating pattern (because 640/256 = 2.5).

Unfortunately reading from main memory doesn't work at this time, so the VGA
buffer runs out of bytes and asserts the empty signal, which the user sees
as a 502 error code on the 7-seg. The LPDDR RAM accepts a read command from
the VGA buffer (we can tell because the cmd_empty signal goes low), but data
never arrives on the read bus.

The 256-byte buffer is constructed using a 9Kib block RAM on the FPGA board.
The Spartan 6 has two of these 9Kib blocks, and each has two ports. We use
one block of RAM, and one port is for writing 32-bit chunks that come back
from main memory and the other is for reading 8-bit colors and handing them
to the VGA display controller.