Fix reading from RAM, VGA buffer works.
The problem was the fact that calibration of the LPDDR RAM
takes around 150,000 cycles, which meant that the VGA buffer
would send a read request to the RAM way before it was ready
to accept transactions.

We will need a boot phase before anything starts executing.
Right now it is reading garbage data from the RAM. The boot
phase needs to wait until the RAM is finished calibrating,
then zero out the video memory. Later on this phase will also
include reading from the micro SD card.