Adds boot stage of pixel writer.
The pixel writer module will be used by the CPU to execute
PIXEL instructions. It is also responsible for zeroing out
the graphics memory during the boot up stage. Currently
only the zeroing out part is programmed in - and there is a
problem with either the reading or writing to graphics memory,
because the screen doesn't show the correct color. It currently
looks like it might be showing uninitialized data. Neither the
overflow, underflow, or error flags ever go high, and the
transactions to and from memory seem to go through. I will
have to write a small proof of concept to isolate and fix the
bug before moving on.