Fixes RAM unitialized memory bug and finishes pixel writer.
The bug was caused because the write mask was the opposite of what
it should have been. I thought the mask was going to be ANDed with
the write data so the bytes would be written only if the mask was
high. It turns out the opposite is true, and the bytes are *masked*
(not written) if the mask is high.

The pixel writer now has an interface for writing individual pixels
once it has finished clearing the screen. The external module sends
a pixel_en signal, an RGB color, and an (x, y) coordinate. The
current build should draw a white pixel in the center of the screen.
Unfortunately it looks like two cyan pixels instead - this may
be a bug in the VGA display or VGA buffer modules.