Makes VGA buffer more robust.
The SD card, when inserted, was causing the VGA buffer's memory
request to stall - that is, the mem_rd_empty flag would never
go low. This essentially froze the VGA buffer's memory contents,
so the screen would just repeat one line from video memory over
and over. To fix this, I have added a timer so that when the VGA
buffer is in the WAIT state, it can time out and return to the
idle state after not receiving data for a certain number of
cycles.

I have additionally changed the inner workings of the VGA buffer
to use the y_coord input to decide which line to fetch next
from memory, rathe than using a rolling buffer with a pointer
to valid memory and a length that represented the number of
valid bytes.

There seems to be a bug where, every 4 pixels or so, the VGA
controller displays the pixel from the line above it, but only
for the first pass in drawing a multi-pixel screen line. This
is likely due to an off-by-one error in the new code for the VGA
buffer, which I will have to fix.