Fixes VGA buffer.
Fixes the bug introduced by the previous commit, in which
columns of VGA pixels would be shifted down by 1 at regular
intervals across the screen. Also fixes a bug from prior
to that commit, in which the display was shifted to the right
by several Consolite pixels, so that the far right columns
wrapped around to the left.