Implements program loading via UART.
After the RAM is calibrated (~1.5 milliseconds after boot), the
user can send 64 KiB of data to the device from their computer
with something like:

cat my_binary > /dev/ttyACM0

where my_binary is a 64 KiB file (or larger, but data after 64 KiB
will be ignored) and /dev/ttyACM0 is the path to the device. I could
not get my computer to output more than 19200 baud (bits/sec), so it
takes about 27 seconds to load a program. You do get progress
information on the 7-seg in the form of a counter while the program
is loading.

I have read the data from the RAM after the UART loader has finished,
and the data appears to be correct. The next step will be to write
the control unit, data cache, and instruction cache, and then hook
everything together.