Implements basic instruction cache.
There is currently a bug - if the data cache overwrites an instruction,
it will not be invalidated and the cache may use the stale instruction
instead. Additionally, only one instruction is cached at a time. In
general, the instruction pointer will change every time the current
instruction is used, and the next instruction will then be fetched.
In the future I will fix the stale instruction bug and cache more
instructions at a time.