Adds processing unit that executes instructions.
At this point about 42% of the board area is taken up, and it only meets
timing constraints (at a max of 104 MHz) if you change the project settings
to optimize for timing. So that will need to be addressed; as of now the
implementation is pretty naive.

The DIV instruction is not supported, because a divider is too large
and slows the maximum clock rate to about 20 MHz. I will probably hand-roll
a divider that divides over the course of multiple clock cycles. Divide
will certainly be an expensive instruction to execute.

The data cache is not currently working, so only programs that don't use
main memory will run. And I am not certain that all instructions besides
those that hit the data cache work, because I have not tested them all.
However I was able to draw a rectangle, so putting pixels and conditional
looping appear to work.

There is an issue with the VGA controller still (possibly the pixel writer
or the VGA buffer). The first four pixels on the left side of the screen
are cut off, and the first four pixels on the right side of the screen
are flickery garbage. There must be an off-by-one-word error somewhere.