Fixes data cache issues.
There was a misspelling of a net entering into the processor module
that was lost among the other warnings. With that fixed, and a few
additional tweaks to data_cache.v, it seems like it's mostly working.
The RET instruction still doesn't work correctly, however, so that
still needs to be fixed. In addition the screen is still shifted to
the left by 4 pixels.