Speeds up SDC load, adds program selector.
The user can now use the 8 on board switches to select which program
to load on power up. If the switches are all 0s, the first program is
loaded. If the switches are in the 00000001 configuration, the program
at a byte offset of 64KiB is loaded, and so on.

The SD card load has been sped up by increasing the clock rate to 10MHz
from 400KHz after software reset.